refactor(auto_learn): extract correction algorithm into submodule
Move the pure correction-detection functions (edit distance, tokenization,
candidate scoring/confidence, anchor location, span diff/alignment, and the
baseline-capture wrappers) out of the 2.1k-line auto_learn.rs into
api/auto_learn/correction.rs. The data structs, tuning constants, and
StableTextGate stay in the parent, so the moved functions borrow them as a
child module with no field-visibility churn; they're re-exported via a glob so
record_candidate and the test module call them unchanged. Behavior-preserving
(all 222 tests pass); auto_learn.rs drops from ~2160 to ~1640 lines.
